Get rid of integers module
